+++ src/services/EventService.ts
@@ -2,8 +2,8 @@ import dayjs from 'dayjs';
 import { PostBaseResponseDto } from '../interfaces/common/PostBaseResponseDto';
 import { EventCreateDto } from '../interfaces/event/EventCreateDto';
 import { EventCreateResponseDto } from '../interfaces/event/EventCreateResponseDto';
-import { EventUpdateResponseDto } from '../interfaces/event/EventCreateResponseDto copy';
 import { EventUpdateDto } from '../interfaces/event/EventUpdateDto';
+import { EventUpdateResponseDto } from '../interfaces/event/EventUpdateResponseDto';
 import Event from '../models/Event';
 import checkObjectIdValidation from '../modules/checkObjectIdValidation';
 import checkValidUtils from '../modules/checkValidUtils';
